added jwt with RefreshToken; added kafka configuration; added billing service; added notification service; added oreder-service;

+++ auth-service/src/main/java/com/otushomework/authservice/model/User.java
@@ -8,16 +8,15 @@ public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
-
+    private Long id;
     private String username;
 
     // Getters and Setters
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
